Refactored admin.py into separate files, linting
